server/requirements.txt
Flask
python-dotenv
sendgrid
